Got everything working for me again
I still had your build path in my .classpath file and Eclipse really
didn't like that but it took me a while to figure that out.  I don't
think I changed anything else.  Also it turns out that the ARPACK
doesn't like to exceed a certain number of iterations on their
diagonalization so there's a limit to how big we can make the matrices
we feed to it.  I came up with about 1200 elements depending on the
domain and the height of the well