More Main Testing
Everything seems to be in working order.  I was having problems earlier
this week with scaling the charge density but then I remembered that we
have to include the dopant density.  The solutions for the ground state
with a rough approximation of how many electrons are in each energy
level is about 5.8 meV which seems close to me.